Separated message parser into market parser and user parser.
