+++ binancefutures-api/src/main/java/io/contek/invoker/binancefutures/api/websocket/WebSocketApi.java
@@ -6,6 +6,7 @@ import io.contek.invoker.commons.actor.ratelimit.RateLimitQuota;
 import io.contek.invoker.commons.websocket.AnyWebSocketMessage;
 import io.contek.invoker.commons.websocket.BaseWebSocketApi;
 import io.contek.invoker.commons.websocket.IWebSocketAuthenticator;
+import io.contek.invoker.commons.websocket.IWebSocketMessageParser;
 
 import javax.annotation.concurrent.ThreadSafe;
 
@@ -14,8 +15,8 @@ public abstract class WebSocketApi extends BaseWebSocketApi {
 
   private final WebSocketRequestIdGenerator requestIdGenerator = new WebSocketRequestIdGenerator();
 
-  protected WebSocketApi(IActor actor) {
-    super(actor, WebSocketMessageParser.getInstance(), IWebSocketAuthenticator.noOp());
+  protected WebSocketApi(IActor actor, IWebSocketMessageParser parser) {
+    super(actor, parser, IWebSocketAuthenticator.noOp());
   }
 
   public final WebSocketRequestIdGenerator getRequestIdGenerator() {
