+++ binancefutures-api/src/main/java/io/contek/invoker/binancefutures/api/websocket/user/OrderUpdateChannel.java
@@ -1,7 +1,5 @@
 package io.contek.invoker.binancefutures.api.websocket.user;
 
-import io.contek.invoker.binancefutures.api.websocket.UserWebSocketChannel;
-
 public final class OrderUpdateChannel extends UserWebSocketChannel<OrderUpdateEvent> {
 
   @Override
